added alert dialog for sorting preferences, implemented sharedpreferences, condenced api calls and api results classes, changes to strings and arrays
