+++ app/src/main/java/com/charlesrowland/popularmovies/MainActivity.java
@@ -1,20 +1,32 @@
 package com.charlesrowland.popularmovies;
 
 
+import android.app.AlertDialog;
+import android.app.Dialog;
+import android.app.ProgressDialog;
 import android.content.Context;
+import android.content.DialogInterface;
+import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
+import android.preference.PreferenceManager;
 import android.support.constraint.ConstraintLayout;
 import android.support.v4.widget.SwipeRefreshLayout;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.GridLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.util.Log;
+import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
+import android.widget.CheckBox;
+import android.widget.EditText;
+import android.widget.RadioButton;
+import android.widget.RadioGroup;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import java.util.List;
 import java.util.Objects;
@@ -43,6 +55,9 @@ public class MainActivity extends AppCompatActivity {
     private GridLayoutManager gridLayoutManager;
     private SwipeRefreshLayout swipeRefreshLayout;
     private TextView noInternetTextView;
+    private View alertLayout;
+    private SharedPreferences mPreferences;
+    private SharedPreferences.Editor mEditor;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -53,6 +68,9 @@ public class MainActivity extends AppCompatActivity {
 
         noInternetTextView = findViewById(R.id.internet_out_message);
 
+        swipeRefreshLayout = findViewById(R.id.swipe_refresh);
+        setSwipeRefreshLayout();
+
         // separate method for RecyclerView setup
         setupRecyclerView();
 
@@ -62,8 +80,6 @@ public class MainActivity extends AppCompatActivity {
         } else {
             showNoNetwork();
         }
-
-        //TODO: setup the swipe to refresh shit
     }
 
     private void showPosters() {
@@ -77,7 +93,12 @@ public class MainActivity extends AppCompatActivity {
     }
 
     private void setSwipeRefreshLayout() {
-        swipeRefreshLayout = findViewById(R.id.swipe_refresh);
+        swipeRefreshLayout.setOnRefreshListener(new SwipeRefreshLayout.OnRefreshListener() {
+            @Override
+            public void onRefresh() {
+                buildAdapter();
+            }
+        });
     }
 
     private void setupRecyclerView() {
@@ -91,15 +112,28 @@ public class MainActivity extends AppCompatActivity {
     }
 
     private void buildAdapter() {
+        swipeRefreshLayout.setRefreshing(false);
+
+        Call<MovieSortingWrapper> call;
         ApiInterface api = ApiClient.getRetrofit().create(ApiInterface.class);
-        Call<MovieSortingWrapper> call = api.getPopularMovies();
+
+        Log.i(TAG, "buildAdapter: orderby: " + getSharedPreferenceOrderbyValue());
+
+        if (getSharedPreferenceOrderbyValue().equals(getResources().getString(R.string.settings_order_by_default))) {
+            call = api.getPopularMovies();
+        } else {
+            call = api.getTopRatedMovies();
+        }
+        
         call.enqueue(new Callback<MovieSortingWrapper>() {
             @Override
             public void onResponse(Call<MovieSortingWrapper> call, Response<MovieSortingWrapper> response) {
+                Log.i(TAG, "onResponse: hello?");
                 MovieSortingWrapper movie = response.body();
                 mAdapter = new MovieAdapter(results);
                 mAdapter.setData(movie.getResults());
                 mMoviePosterRecyclerView.setAdapter(mAdapter);
+                mAdapter.notifyDataSetChanged();
             }
 
             @Override
@@ -107,6 +141,13 @@ public class MainActivity extends AppCompatActivity {
                 showNoNetwork();
             }
         });
+    }
+
+    private String getSharedPreferenceOrderbyValue() {
+        mPreferences = PreferenceManager.getDefaultSharedPreferences(this);
+        String orderby_key = getResources().getString(R.string.settings_order_by_key);
+
+        return mPreferences.getString(orderby_key, getResources().getString(R.string.settings_order_by_default));
     }
 
     private boolean checkNetworkStatus() {
@@ -119,7 +160,6 @@ public class MainActivity extends AppCompatActivity {
         if (networkInfo != null && networkInfo.isConnected()) {
             hasNetworkConn = true;
         }
-        Log.i(TAG, "checkNetworkStatus: is there a network connection? ");
         return hasNetworkConn;
     }
 
@@ -154,6 +194,8 @@ public class MainActivity extends AppCompatActivity {
     public boolean onOptionsItemSelected(MenuItem item){
         switch(item.getItemId()){
             case R.id.action_show_options:
+                Log.i(TAG, "onOptionsItemSelected: CLICKED");
+                showSortOptionsDialog();
                 return true;
         }
 
@@ -168,5 +210,59 @@ public class MainActivity extends AppCompatActivity {
     @Override
     public void onDestroy() {
         super.onDestroy();
+    }
+
+    public void showSortOptionsDialog() {
+        LayoutInflater inflater = getLayoutInflater();
+        alertLayout = inflater.inflate(R.layout.settings_dialog, null);
+
+        final RadioGroup sortGroup = alertLayout.findViewById(R.id.movieSortOrder);
+        RadioButton popularity_button = alertLayout.findViewById(R.id.popularity);
+        RadioButton toprated_button = alertLayout.findViewById(R.id.top_rated);
+
+        if (getSharedPreferenceOrderbyValue().equals(getResources().getString(R.string.settings_order_by_default))) {
+            popularity_button.setChecked(true);
+        } else {
+            toprated_button.setChecked(true);
+        }
+
+        AlertDialog.Builder alert = new AlertDialog.Builder(this);
+        alert.setView(alertLayout);
+        alert.setMessage("Options");
+        alert.setCancelable(false);
+
+        alert.setNegativeButton("Cancel", new DialogInterface.OnClickListener() {
+
+            @Override
+            public void onClick(DialogInterface dialog, int which) {
+
+            }
+        });
+
+        alert.setPositiveButton("Done", new DialogInterface.OnClickListener() {
+
+            @Override
+            public void onClick(DialogInterface dialog, int which) {
+                String sortOrder = getSelectedSortOrder(sortGroup);
+
+                if (!getSharedPreferenceOrderbyValue().equals(sortOrder)) {
+                    mEditor = mPreferences.edit();
+                    mEditor.putString(getResources().getString(R.string.settings_order_by_key), sortOrder);
+                    mEditor.commit();
+                    buildAdapter();
+                }
+            }
+        });
+
+        AlertDialog dialog = alert.create();
+        dialog.show();
+    }
+
+    private String getSelectedSortOrder(RadioGroup sortGroup) {
+        int radioButtonID = sortGroup.getCheckedRadioButtonId();
+        View radioButton = sortGroup.findViewById(radioButtonID);
+        int idx = sortGroup.indexOfChild(radioButton);
+        RadioButton r = (RadioButton)  sortGroup.getChildAt(idx);
+        return r.getText().toString().toLowerCase();
     }
 }
