--- app/src/main/java/com/charlesrowland/popularmovies/model/MovieCreditsCast.java
@@ -1,103 +1,0 @@
-package com.charlesrowland.popularmovies.model;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class MovieCreditsCast {
-
-    @SerializedName("cast_id")
-    @Expose
-    private Integer cast_id;
-
-    @SerializedName("character")
-    @Expose
-    private String character;
-
-    @SerializedName("credit_id")
-    @Expose
-    private String credit_id;
-
-    @SerializedName("gender")
-    @Expose
-    private Integer gender;
-
-    @SerializedName("id")
-    @Expose
-    private Integer id;
-
-    @SerializedName("name")
-    @Expose
-    private String name;
-
-    @SerializedName("order")
-    @Expose
-    private Integer order;
-
-    @SerializedName("profile_path")
-    @Expose
-    private String profile_path;
-
-    public Integer getCast_id() {
-        return cast_id;
-    }
-
-    public void setCast_id(Integer cast_id) {
-        this.cast_id = cast_id;
-    }
-
-    public String getCharacter() {
-        return character;
-    }
-
-    public void setCharacter(String character) {
-        this.character = character;
-    }
-
-    public String getCredit_id() {
-        return credit_id;
-    }
-
-    public void setCredit_id(String credit_id) {
-        this.credit_id = credit_id;
-    }
-
-    public Integer getGender() {
-        return gender;
-    }
-
-    public void setGender(Integer gender) {
-        this.gender = gender;
-    }
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public Integer getOrder() {
-        return order;
-    }
-
-    public void setOrder(Integer order) {
-        this.order = order;
-    }
-
-    public String getProfile_path() {
-        return profile_path;
-    }
-
-    public void setProfile_path(String profile_path) {
-        this.profile_path = profile_path;
-    }
-}
