--- app/src/main/java/com/charlesrowland/popularmovies/model/MovieDetailsResult.java
@@ -1,214 +1,0 @@
-package com.charlesrowland.popularmovies.model;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-import java.util.List;
-
-public class MovieDetailsResult {
-
-    @SerializedName("backdrop_path")
-    @Expose
-    private String backdrop_path;
-
-    @SerializedName("budget")
-    @Expose
-    private Integer budget;
-
-    @SerializedName("genres")
-    @Expose
-    private List<MovieGenreResult> genres = null;
-
-    @SerializedName("homepage")
-    @Expose
-    private String homepage;
-
-    @SerializedName("id")
-    @Expose
-    private Integer id;
-
-    @SerializedName("imdb_id")
-    @Expose
-    private String imdb_id;
-
-    @SerializedName("original_language")
-    @Expose
-    private String originalLanguage;
-
-    @SerializedName("original_title")
-    @Expose
-    private String originalTitle;
-
-    @SerializedName("overview")
-    @Expose
-    private String overview;
-
-    @SerializedName("popularity")
-    @Expose
-    private Double popularity;
-
-    @SerializedName("poster_path")
-    @Expose
-    private String posterPath;
-
-    @SerializedName("release_date")
-    @Expose
-    private String releaseDate;
-
-    @SerializedName("revenue")
-    @Expose
-    private Integer revenue;
-
-    @SerializedName("runtime")
-    @Expose
-    private Integer runtime;
-
-    @SerializedName("tagline")
-    @Expose
-    private String tagline;
-
-    @SerializedName("vote_average")
-    @Expose
-    private Double voteAverage;
-
-    @SerializedName("vote_count")
-    @Expose
-    private Integer voteCount;
-
-
-    public String getBackdrop_path() {
-        return backdrop_path;
-    }
-
-    public void setBackdrop_path(String backdrop_path) {
-        this.backdrop_path = backdrop_path;
-    }
-
-    public Integer getBudget() {
-        return budget;
-    }
-
-    public void setBudget(Integer budget) {
-        this.budget = budget;
-    }
-
-    public List<MovieGenreResult> getGenres() {
-        return genres;
-    }
-
-    public void setGenres(List<MovieGenreResult> genres) {
-        this.genres = genres;
-    }
-
-    public String getHomepage() {
-        return homepage;
-    }
-
-    public void setHomepage(String homepage) {
-        this.homepage = homepage;
-    }
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getImdb_id() {
-        return imdb_id;
-    }
-
-    public void setImdb_id(String imdb_id) {
-        this.imdb_id = imdb_id;
-    }
-
-    public String getOriginalLanguage() {
-        return originalLanguage;
-    }
-
-    public void setOriginalLanguage(String originalLanguage) {
-        this.originalLanguage = originalLanguage;
-    }
-
-    public String getOriginalTitle() {
-        return originalTitle;
-    }
-
-    public void setOriginalTitle(String originalTitle) {
-        this.originalTitle = originalTitle;
-    }
-
-    public String getOverview() {
-        return overview;
-    }
-
-    public void setOverview(String overview) {
-        this.overview = overview;
-    }
-
-    public Double getPopularity() {
-        return popularity;
-    }
-
-    public void setPopularity(Double popularity) {
-        this.popularity = popularity;
-    }
-
-    public String getPosterPath() {
-        return posterPath;
-    }
-
-    public void setPosterPath(String posterPath) {
-        this.posterPath = posterPath;
-    }
-
-    public String getReleaseDate() {
-        return releaseDate;
-    }
-
-    public void setReleaseDate(String releaseDate) {
-        this.releaseDate = releaseDate;
-    }
-
-    public Integer getRevenue() {
-        return revenue;
-    }
-
-    public void setRevenue(Integer revenue) {
-        this.revenue = revenue;
-    }
-
-    public Integer getRuntime() {
-        return runtime;
-    }
-
-    public void setRuntime(Integer runtime) {
-        this.runtime = runtime;
-    }
-
-    public String getTagline() {
-        return tagline;
-    }
-
-    public void setTagline(String tagline) {
-        this.tagline = tagline;
-    }
-
-    public Double getVoteAverage() {
-        return voteAverage;
-    }
-
-    public void setVoteAverage(Double voteAverage) {
-        this.voteAverage = voteAverage;
-    }
-
-    public Integer getVoteCount() {
-        return voteCount;
-    }
-
-    public void setVoteCount(Integer voteCount) {
-        this.voteCount = voteCount;
-    }
-}
