+++ app/src/main/res/drawable/ic_checkmark.xml
@@ -1,0 +1,9 @@
+<vector
+    android:height="48dp"
+    android:tint="#3B8730"
+    android:viewportHeight="24.0"
+    android:viewportWidth="24.0"
+    android:width="48dp"
+    xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#FF000000" android:pathData="M9,16.17L4.83,12l-1.42,1.41L9,19 21,7l-1.41,-1.41z"/>
+</vector>
