+++ app/src/main/res/values/arrays.xml
@@ -1,4 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
+    <string-array name="settings_order_by_labels">
+        <item>@string/settings_order_by_popularity_label</item>
+        <item>@string/settings_order_by_toprated_label</item>
+    </string-array>
 
+    <string-array name="settings_order_by_values">
+        <item>@string/settings_order_by_popularity_value</item>
+        <item>@string/settings_order_by_toprated_value</item>
+    </string-array>
 </resources>
