+++ app/src/main/res/values/strings.xml
@@ -6,12 +6,15 @@
     <string name="no_internet">Ah nuts, it looks like there is a problem with your internet connection. You should probably get that fixed.</string>
 
     <string name="menu_title_options">Options</string>
+    <string name="menu_dialog_title">Options</string>
 
-    <string name="settings_order_by_label">Order Movies By</string>
     <string name="settings_order_by_key" translatable="false">order_by</string>
-    <string name="settings_order_by_default" translatable="false">popularity.desc</string>
-    <!--SORTING VALUES:
-    popularity.asc, popularity.desc, release_date.asc, release_date.desc, revenue.asc, revenue.desc, primary_release_date.asc, primary_release_date.desc, original_title.asc, original_title.desc, vote_average.asc, vote_average.desc, vote_count.asc, vote_count.desc-->
+    <string name="settings_order_by_default" translatable="false">popularity</string>
 
+    <string name="settings_order_by_popularity_label">Popularity</string>
+    <string name="settings_order_by_popularity_value" translatable="false">popularity</string>
+
+    <string name="settings_order_by_toprated_label">Top Rated</string>
+    <string name="settings_order_by_toprated_value" translatable="false">top rated</string>
 
 </resources>
